backend/requirements.txt
Django==5.2.11
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
python-decouple==3.8
requests==2.32.3
psycopg2-binary==2.9.10
gunicorn==21.2.0

# For production (optional)
# redis==5.0.1
# celery==5.3.4
